Fix compatibility issue for ColumnarBatchScanExec
