+++ shims/spark311/src/main/scala/com/intel/oap/execution/BatchScanExecFake.scala
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2020 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.intel.oap.sql.execution
+
+import org.apache.spark.sql.execution.datasources.v2.BatchScanExec
+
+public class BatchScanExecFake(output: Seq[AttributeReference], @transient scan: Scan,
+                               runtimeFilters: Seq[Expression])
+  extends BatchScanExec(output, scan) {
+}
